Make token-balance-cache into library.
