+++ package.json
@@ -14,11 +14,21 @@
   "keywords": [],
   "workspaces": [
     "packages/common-common",
+    "packages/token-balance-cache",
     "packages/commonwealth"
   ],
   "dependencies": {
+    "@cosmjs/stargate": "^0.26.1",
+    "@cosmjs/tendermint-rpc": "^0.26.1",
+    "@types/moment": "^2.13.0",
+    "@solana/web3.js": "^1.30.2",
+    "axios": "^0.21.2",
     "ethers": "5.4.3",
-    "typescript": "4.3.4"
+    "moment": "^2.23.0",
+    "pg": "^8.7.1",
+    "typescript": "4.3.4",
+    "web3": "1.5.2",
+    "web3-utils": "1.5.2"
   },
   "devDependencies": {
   }
