+++ packages/common-common/package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@nomiclabs/hardhat-ethers": "^2.0.2",
     "@typechain/ethers-v5": "^6.0.0",
+    "async-rwlock": "^1.1.1",
     "hardhat": "^2.3.0",
     "hardhat-typechain": "^0.3.5",
     "ts-generator": "^0.1.1",
