+++ packages/commonwealth/package.json
@@ -66,8 +66,6 @@
     "@commonwealth/chain-events": "^0.13.3",
     "@cosmjs/amino": "^0.26.1",
     "@cosmjs/proto-signing": "^0.26.1",
-    "@cosmjs/stargate": "^0.26.1",
-    "@cosmjs/tendermint-rpc": "^0.26.1",
     "@ethersproject/abi": "^5.0.1",
     "@ethersproject/address": "^5.0.1",
     "@ethersproject/bytes": "^5.0.5",
@@ -97,7 +95,6 @@
     "@socket.io/admin-ui": "^0.2.0",
     "@socket.io/postgres-adapter": "^0.1.1",
     "@socket.io/redis-adapter": "^7.1.0",
-    "@solana/web3.js": "^1.30.2",
     "@terra-money/terra.js": "^3.1.3",
     "@types/bluebird": "^3.5.36",
     "@types/bn.js": "^4.11.6",
@@ -112,7 +109,6 @@
     "@types/marked": "^4.0.1",
     "@types/mithril": "^2.0.7",
     "@types/mocha": "^7.0.2",
-    "@types/moment": "^2.13.0",
     "@types/node-sass": "^4.11.2",
     "@types/passport": "^1.0.5",
     "@types/passport-jwt": "^3.0.3",
@@ -122,10 +118,8 @@
     "@walletconnect/web3-provider": "^1.7.0",
     "amqplib": "^0.8.0",
     "assert": "^2.0.0",
-    "async-rwlock": "^1.1.1",
     "autoprefixer": "^10.2.5",
     "aws-sdk": "^2.814.0",
-    "axios": "^0.21.2",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.4",
     "bcrypt-nodejs": "^0.0.3",
@@ -179,7 +173,6 @@
     "mithril-infinite": "^1.3.0",
     "mixpanel": "^0.14.0",
     "mixpanel-browser": "^2.45.0",
-    "moment": "^2.23.0",
     "moment-locales-webpack-plugin": "^1.1.0",
     "morgan": "^1.9.1",
     "near-api-js": "^0.42.0",
@@ -195,7 +188,6 @@
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "passport-magic": "^1.0.0",
-    "pg": "^8.7.1",
     "postcss": "^8.2.13",
     "postcss-import": "^14.0.1",
     "postcss-loader": "4.2.0",
@@ -231,7 +223,6 @@
     "tweetnacl": "^1.0.1",
     "underscore": "^1.13.1",
     "velocity": "^0.7.2",
-    "web3": "1.5.2",
     "webpack": "^4.36.0",
     "webpack-bundle-analyzer": "^4.4.0",
     "webpack-cli": "^3.1.2",
