+++ packages/commonwealth/server-test.ts
@@ -15,6 +15,8 @@ import {
   NotificationCategories,
   ChainType,
 } from 'common-common/src/types';
+import TokenBalanceCache from 'token-balance-cache/src/index';
+import TokenBalanceProvider from 'token-balance-cache/src/provider';
 
 import {ROLLBAR_SERVER_TOKEN, SESSION_SECRET} from './server/config';
 import setupAPI from './server/router'; // performance note: this takes 15 seconds
@@ -22,7 +24,6 @@ import setupPassport from './server/passport';
 import models from './server/database';
 import ViewCountCache from './server/util/viewCountCache';
 import IdentityFetchCache from './server/util/identityFetchCache';
-import TokenBalanceCache, { TokenBalanceProvider } from './server/util/tokenBalanceCache';
 import BanCache from './server/util/banCheckCache';
 import setupErrorHandlers from './server/scripts/setupErrorHandlers';
 import RuleCache from './server/util/rules/ruleCache';
@@ -55,7 +56,6 @@ class MockTokenBalanceProvider extends TokenBalanceProvider {
 
 const mockTokenBalanceProvider = new MockTokenBalanceProvider();
 const tokenBalanceCache = new TokenBalanceCache(
-  models,
   0,
   0,
   mockTokenBalanceProvider
