+++ packages/commonwealth/server/models/chain_node.ts
@@ -1,5 +1,6 @@
 import * as Sequelize from 'sequelize'; // must use "* as" to avoid scope errors
 import { DataTypes } from 'sequelize';
+import { BalanceType } from 'common-common/src/types';
 import { ModelStatic, ModelInstance } from './types';
 
 export type ChainNodeAttributes = {
@@ -8,6 +9,9 @@ export type ChainNodeAttributes = {
   eth_chain_id?: number;
   alt_wallet_url?: string;
   private_url?: string;
+  balance_type?: BalanceType;
+  name?: string;
+  description?: string;
 }
 
 export type ChainNodeInstance = ModelInstance<ChainNodeAttributes> & {
@@ -28,6 +32,9 @@ export default (
       eth_chain_id: { type: dataTypes.INTEGER, allowNull: true },
       alt_wallet_url: { type: dataTypes.STRING, allowNull: true },
       private_url: { type: dataTypes.STRING, allowNull: true },
+      balance_type: { type: dataTypes.STRING, allowNull: true },
+      name: { type: dataTypes.STRING, allowNull: true },
+      description: { type: dataTypes.TEXT, allowNull: true },
     },
     {
       tableName: 'ChainNodes',
