+++ packages/commonwealth/server/util/banCheckCache.ts
@@ -1,9 +1,10 @@
 // Helper function to look up a scope, i.e. a chain XOR community.
 // If a community is found, also check that the user is allowed to see it.
 import { Op } from 'sequelize';
+import JobRunner from 'common-common/src/cacheJobRunner';
+
 import { DB } from '../database';
 
-import JobRunner from './cacheJobRunner';
 
 export const BanErrors = {
   NoAddress: 'Address not found',
