+++ packages/commonwealth/server/util/rules/ruleCache.ts
@@ -1,5 +1,5 @@
 import moment from 'moment';
-import JobRunner from '../cacheJobRunner';
+import JobRunner from 'common-common/src/cacheJobRunner';
 
 import { factory, formatFilename } from 'common-common/src/logging';
 const log = factory.getLogger(formatFilename(__filename));
