+++ packages/commonwealth/test/tsconfig.node.json
@@ -10,6 +10,8 @@
     "paths": {
       "common-common": ["../../../packages/common-common"],
       "common-common/*": ["../../../packages/common-common/*"],
+      "token-balance-cache": ["../../../packages/token-balance-cache"],
+      "token-balance-cache/*": ["../../../packages/token-balance-cache/*"],
       "*": ["../*", "../shared/*", "../client/scripts/*"]
     },
     "resolveJsonModule": true,
