+++ packages/commonwealth/test/util/modelUtils.ts
@@ -8,10 +8,10 @@ import { signTypedData, SignTypedDataVersion } from '@metamask/eth-sig-util';
 import { Keyring } from '@polkadot/api';
 import { stringToU8a, u8aToHex } from '@polkadot/util';
 import { factory, formatFilename } from 'common-common/src/logging';
+import TokenBalanceProvider from 'token-balance-cache/src/provider';
 import app from '../../server-test';
 import models from '../../server/database';
 import { Permission } from '../../server/models/role';
-import { TokenBalanceProvider } from '../../server/util/tokenBalanceCache';
 import { constructTypedMessage } from '../../shared/adapters/chain/ethereum/keys';
 
 const log = factory.getLogger(formatFilename(__filename));
