+++ packages/commonwealth/tsconfig.json
@@ -11,6 +11,8 @@
     "paths": {
       "common-common": ["../../packages/common-common"],
       "common-common/*": ["../../packages/common-common/*"],
+      "token-balance-cache": ["../../packages/token-balance-cache"],
+      "token-balance-cache/*": ["../../packages/token-balance-cache/*"],
       "*": ["./*", "shared/*", "client/scripts/*"]
     },
     "resolveJsonModule": true,
