+++ packages/commonwealth/webpack/webpack.common.js
@@ -84,7 +84,8 @@ module.exports = {
       '../node_modules', // global node modules
     ],
     alias: {
-      "common-common": path.resolve(__dirname, '../../common-common')
+      "common-common": path.resolve(__dirname, '../../common-common'),
+      "token-balance-cache": path.resolve(__dirname, '../../token-balance-cache'),
     }
   },
   module: {
@@ -119,6 +120,7 @@ module.exports = {
           path.resolve(__dirname, '../client'),
           path.resolve(__dirname, '../shared'),
           path.resolve(__dirname, '../../common-common'),
+          path.resolve(__dirname, '../../token-balance-cache'),
         ],
         loader: 'esbuild-loader',
         options: {
